Update: functionality of get_student_by_id
Previously, the function would parse the json object contained in the
file and extract the json array under the 'students' key, as a Python
list. The student_id passed as an argument in the URI, was directly used to access the index within that list. This was abug because lists are zero-indexed. This was therefore changed to have the function look for the json student object that matches the student id argument within the list, and return it