Update: how students dict is initialised
Previously, the global variable was initialised to a NoneType. This made
the program bulky and a few lines of code were necessary to handle
different scenarios. This has been changed to initialising it to a
Python dict, where the 'size' key is set to 0 and the 'students' key is
set to an empty list