Update: how student is added in add_student route
Previously, with how the global students variable was being initialised
to a NoneType, extra code was required to handle different scenarios
especially when students was None. This has been changed and now there
are no checks needed. The size variable is simply incremented, while the
new student dict object is appended to the students list